+++ src/test/java/org/springframework/beans/factory/bsafb/BSAFBIUserDao.java
@@ -1,0 +1,12 @@
+package org.springframework.beans.factory.bsafb;
+
+/**
+ * BSAFB 是 bean-scope-and-factory-bean 分支名称的首字母
+ *
+ * @author: DoubleW2w
+ * @date: 2024/10/23
+ * @project: sbs-small-spring
+ */
+public interface BSAFBIUserDao {
+  String queryUserName(String uId);
+}